refactor(delta): extract msgNorm + findSharedPrefix into shared helper
Both server/index.js (delta write decision) and scripts/migrate-to-delta.js
carried inline copies of msgNorm and findSharedPrefix-style logic. The
migration script's comment explicitly flagged it as "Copied verbatim from
server/index.js — must stay in sync"; that drift risk goes away by giving
both files a single source of truth in server/delta-helpers.js.

The helper exposes two variants:
  findSharedPrefix         — full prevMsgs array (live server uses this)
  findSharedPrefixFromLast — last msg + count only (memory-minimal, used
                             by the migration script under the alias
                             canDelta to keep its existing call sites).

No behaviour change. All existing tests still pass; live server delta
write and migration script dry-run produce identical output.
